ai_agent/requirements.txt
web3>=7.0.0
python-dotenv>=1.0
loguru>=0.7.0
schedule>=1.2.1
requests>=2.32.3
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
openai>=1.0.0
anthropic>=0.18.0
